Broke apart the modules.rs file into different files.
