+++ server/src/types/stations/reducers.rs
@@ -5,7 +5,7 @@ use crate::types::{
     items::{ GetItemDefinitionRowOptionById, ItemDefinitionId },
     players::{ GetPlayerRowOptionById, UpdatePlayerRowById },
     ships::{ utility::*, * },
-    stations::modules::GetTradingPortModuleRowOptionById,
+
 };
 
 use super::*;
